+++ lib/blazer.rb
@@ -5,6 +5,8 @@ require "safely/core"
 # stdlib
 require "digest/sha2"
 require "json"
+require "securerandom"
+require "tempfile"
 require "yaml"
 
 # modules
@@ -14,6 +16,7 @@ require_relative "blazer/result"
 require_relative "blazer/result_cache"
 require_relative "blazer/run_statement"
 require_relative "blazer/statement"
+require_relative "blazer/tempfile_body"
 
 # adapters
 require_relative "blazer/adapters/base_adapter"
@@ -69,6 +72,7 @@ module Blazer
     attr_accessor :anomaly_checks
     attr_accessor :forecasting
     attr_accessor :async
+    attr_accessor :streaming_csv
     attr_accessor :images
     attr_accessor :override_csp
     attr_accessor :slack_oauth_token
@@ -81,6 +85,7 @@ module Blazer
   self.anomaly_checks = false
   self.forecasting = false
   self.async = false
+  self.streaming_csv = true
   self.images = false
   self.override_csp = false